pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["openclaw_k8s_toggle_operator"]

[project]
name = "openclaw-k8s-toggle-operator"
dynamic = ["version"]
authors = [
  { name="Henning Thie\u00df", email="ht@xomox.cc" },
]
description = "Matrix-controlled Kubernetes deployment toggle operator"
readme = "README.md"
requires-python = ">=3.14"
license-files = ["LICENSE.md"]

keywords = ["matrix", "kubernetes", "k8s", "operator", "deployment", "toggle", "chat-ops", "e2e-encryption"]

# https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 4 - Beta",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",

    "Topic :: System :: Systems Administration",
    "Topic :: Communications :: Chat",

    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
]

dependencies = [
    'aiohttp>=3.9.0',
    'kubernetes>=28.1.0',
    'loguru>=0.7.3',
    'matrix-nio[e2e]>=0.25.0',
    'tabulate>=0.9.0',
]

[project.scripts]
openclaw-k8s-toggle-operator = "openclaw_k8s_toggle_operator.__main__:main"
openclaw-k8s-toggle-operator-conntest = "openclaw_k8s_toggle_operator.__main__:connectortest"

[project.urls]
Homepage = "https://github.com/vroomfondel/openclaw-k8s-toggle-operator"
Repository = "https://github.com/vroomfondel/openclaw-k8s-toggle-operator"
Issues = "https://github.com/vroomfondel/openclaw-k8s-toggle-operator/issues"

[tool.hatch.version]
path = "openclaw_k8s_toggle_operator/__init__.py"

[tool.isort]
profile = "black"
line_length = 120
